Split DropdownNavbarItem into Mobile/Desktop subcomponents
