+++ packages/docusaurus-theme-classic/src/theme/NavbarItem/DropdownNavbarItem/Mobile/styles.module.css
