+++ packages/docusaurus-theme-classic/src/theme/NavbarItem/DropdownNavbarItem/index.tsx
@@ -5,235 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, {
-  useState,
-  useRef,
-  useEffect,
-  type ReactNode,
-  type ComponentProps,
-} from 'react';
-import clsx from 'clsx';
-import {
-  isRegexpStringMatch,
-  useCollapsible,
-  Collapsible,
-} from '@docusaurus/theme-common';
-import {isSamePath, useLocalPathname} from '@docusaurus/theme-common/internal';
-import NavbarNavLink from '@theme/NavbarItem/NavbarNavLink';
-import NavbarItem, {type LinkLikeNavbarItemProps} from '@theme/NavbarItem';
-import type {
-  DesktopOrMobileNavBarItemProps,
-  Props,
-} from '@theme/NavbarItem/DropdownNavbarItem';
-import styles from './styles.module.css';
-
-function isItemActive(
-  item: LinkLikeNavbarItemProps,
-  localPathname: string,
-): boolean {
-  if (isSamePath(item.to, localPathname)) {
-    return true;
-  }
-  if (isRegexpStringMatch(item.activeBaseRegex, localPathname)) {
-    return true;
-  }
-  if (item.activeBasePath && localPathname.startsWith(item.activeBasePath)) {
-    return true;
-  }
-  return false;
-}
-
-function containsActiveItems(
-  items: readonly LinkLikeNavbarItemProps[],
-  localPathname: string,
-): boolean {
-  return items.some((item) => isItemActive(item, localPathname));
-}
-
-function DropdownNavbarItemDesktop({
-  items,
-  position,
-  className,
-  onClick,
-  ...props
-}: DesktopOrMobileNavBarItemProps) {
-  const dropdownRef = useRef<HTMLDivElement>(null);
-  const [showDropdown, setShowDropdown] = useState(false);
-
-  useEffect(() => {
-    const handleClickOutside = (
-      event: MouseEvent | TouchEvent | FocusEvent,
-    ) => {
-      if (
-        !dropdownRef.current ||
-        dropdownRef.current.contains(event.target as Node)
-      ) {
-        return;
-      }
-      setShowDropdown(false);
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    document.addEventListener('touchstart', handleClickOutside);
-    document.addEventListener('focusin', handleClickOutside);
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-      document.removeEventListener('touchstart', handleClickOutside);
-      document.removeEventListener('focusin', handleClickOutside);
-    };
-  }, [dropdownRef]);
-
-  return (
-    <div
-      ref={dropdownRef}
-      className={clsx('navbar__item', 'dropdown', 'dropdown--hoverable', {
-        'dropdown--right': position === 'right',
-        'dropdown--show': showDropdown,
-      })}>
-      <NavbarNavLink
-        aria-haspopup="true"
-        aria-expanded={showDropdown}
-        role="button"
-        // # hash permits to make the <a> tag focusable in case no link target
-        // See https://github.com/facebook/docusaurus/pull/6003
-        // There's probably a better solution though...
-        href={props.to ? undefined : '#'}
-        className={clsx('navbar__link', className)}
-        {...props}
-        onClick={props.to ? undefined : (e) => e.preventDefault()}
-        onKeyDown={(e) => {
-          if (e.key === 'Enter') {
-            e.preventDefault();
-            setShowDropdown(!showDropdown);
-          }
-        }}>
-        {props.children ?? props.label}
-      </NavbarNavLink>
-      <ul className="dropdown__menu">
-        {items.map((childItemProps, i) => (
-          <NavbarItem
-            isDropdownItem
-            activeClassName="dropdown__link--active"
-            {...childItemProps}
-            key={i}
-          />
-        ))}
-      </ul>
-    </div>
-  );
-}
-
-// TODO temp impl
-function CollapseButton({
-  collapsed,
-  onClick,
-}: {
-  collapsed: boolean;
-  onClick: ComponentProps<'button'>['onClick'];
-}) {
-  return (
-    <button
-      // TODO
-      aria-label="TODO"
-      aria-expanded={!collapsed}
-      type="button"
-      className="clean-btn menu__caret"
-      onClick={onClick}
-    />
-  );
-}
-
-function useItemCollapsible({active}: {active: boolean}) {
-  const {collapsed, toggleCollapsed, setCollapsed} = useCollapsible({
-    initialState: () => !active,
-  });
-
-  // Expand if any item active after a navigation
-  useEffect(() => {
-    if (active) {
-      setCollapsed(false);
-    }
-  }, [active, setCollapsed]);
-
-  return {
-    collapsed,
-    toggleCollapsed,
-  };
-}
-
-function DropdownNavbarItemMobile({
-  items,
-  className,
-  position, // Need to destructure position from props so that it doesn't get passed on.
-  onClick,
-  ...props
-}: DesktopOrMobileNavBarItemProps) {
-  const localPathname = useLocalPathname();
-  const isActive = isSamePath(props.to, localPathname);
-  const containsActive = containsActiveItems(items, localPathname);
-
-  const {collapsed, toggleCollapsed} = useItemCollapsible({
-    active: isActive || containsActive,
-  });
-
-  // # hash permits to make the <a> tag focusable in case no link target
-  // See https://github.com/facebook/docusaurus/pull/6003
-  // There's probably a better solution though...
-  const href = props.to ? undefined : '#';
-
-  return (
-    <li
-      className={clsx('menu__list-item', {
-        'menu__list-item--collapsed': collapsed,
-      })}>
-      <div
-        className={clsx('menu__list-item-collapsible', {
-          'menu__list-item-collapsible--active': isActive,
-        })}>
-        <NavbarNavLink
-          role="button"
-          className={clsx(
-            styles.dropdownNavbarItemMobile,
-            'menu__link menu__link--sublist',
-            className,
-          )}
-          href={href}
-          {...props}
-          onClick={(e) => {
-            // Prevent navigation when link is "#"
-            if (href === '#') {
-              e.preventDefault();
-            }
-            // Otherwise we let navigation eventually happen, and/or collapse
-            toggleCollapsed();
-          }}>
-          {props.children ?? props.label}
-        </NavbarNavLink>
-        <CollapseButton
-          collapsed={collapsed}
-          onClick={(e) => {
-            e.preventDefault();
-            toggleCollapsed();
-          }}
-        />
-      </div>
-
-      <Collapsible lazy as="ul" className="menu__list" collapsed={collapsed}>
-        {items.map((childItemProps, i) => (
-          <NavbarItem
-            mobile
-            isDropdownItem
-            onClick={onClick}
-            activeClassName="menu__link--active"
-            {...childItemProps}
-            key={i}
-          />
-        ))}
-      </Collapsible>
-    </li>
-  );
-}
+import React, {type ReactNode} from 'react';
+import DropdownNavbarItemMobile from '@theme/NavbarItem/DropdownNavbarItem/Mobile';
+import DropdownNavbarItemDesktop from '@theme/NavbarItem/DropdownNavbarItem/Desktop';
+import type {Props} from '@theme/NavbarItem/DropdownNavbarItem';
 
 export default function DropdownNavbarItem({
   mobile = false,
